% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjCausDisc-package.R
\docType{package}
\name{AdjCausDisc-package}
\alias{AdjCausDisc}
\alias{AdjCausDisc-package}
\title{AdjCausDisc: Adjacency Matrix Causal Discovery Coefficient Matrix Creator and Visualizer}
\description{
Using an input of Adjacency matrix that is a lower triangular matrix that describes the structure of a DAG, the first function will find the corresponding coefficient matrix based on the Causal Discovery model that find all possible coefficients using spike and slab prior. Then using the DAG structure of the adjacency matrix, the second function will create a visualization of the DAG.
}
\keyword{internal}
