% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefMat_Create_Spike_Slab.R
\name{CoefMat_Create_Spike_Slab}
\alias{CoefMat_Create_Spike_Slab}
\title{CoefMat_Create_Spike_Slab}
\usage{
CoefMat_Create_Spike_Slab(Y, A)
}
\arguments{
\item{Y}{The data the causal discovery adjacency matrix wants to show the connections with.}

\item{A}{The adjacency matrix that shows us the structure of the DAG. The input should be a lower triangular matrix in order for it to be acyclic. The number of rows and number of columns of A should be same as the number of data points in Y.}
}
\value{
A matrix that contains all the coefficients corresponding to specific edge between each entry of the Y data
}
\description{
Uses spike and slab prior to get the corresponding coefficient matrix Bij corresponding to \eqn{Y_{i} = B_{ij}X_{i} + e_{i}}. The coefficients represents the connections from the jth entry of Y to the ith entry of Y.
}
\examples{
Adjacency_matrix = matrix(nrow=3,ncol=3,0)
Adjacency_matrix[,1]=c(0,1,1)
Adjacency_matrix[,2]=c(0,0,1)
Adjacency_matrix[,3]=c(0,0,0)
Y_vector = rnorm(3)
CoefMat_Create_Spike_Slab(Y_vector,Adjacency_matrix)
}
