% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefMat_Create_SSVS.R
\name{CoefMat_Create_SSVS}
\alias{CoefMat_Create_SSVS}
\title{CoefMat_Create_SSVS}
\usage{
CoefMat_Create_SSVS(Y, A)
}
\arguments{
\item{Y}{The data the causal discovery adjacency matrix wants to show the connections with. Y1 = c(y1, ..., yn)}

\item{A}{The adjacency matrix that shows us the structure of the DAG. The input should be a lower triangular matrix where the possible matrix entry and have the number of rows and number of columns same as the number of data points in Y}
}
\value{
A matrix that contains all the coefficients corresponding to specific relationships between each entry of the Y data using Stochastic Search Variable Search to get the values of the coeffecients.
}
\description{
Uses ssvs bayesian model selection to get the corresponding coefficient matrix Bij corresponding to \eqn{Y = B_{ij}X_{i} + e_{i}}. The coefficients represents the connections from the jth entry of Y to the ith entry of Y.
}
\examples{
Adjacency_matrix = matrix(nrow=3,ncol=3,0)
Adjacency_matrix[,1]=c(0,1,1)
Adjacency_matrix[,2]=c(0,0,1)
Adjacency_matrix[,3]=c(0,0,0)
Y_vector = rnorm(3)
CoefMat_Create_SSVS(Y_vector,Adjacency_matrix)
}
