% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjDAG.R
\name{AdjDAG}
\alias{AdjDAG}
\title{AdjDAG}
\usage{
AdjDAG(A)
}
\arguments{
\item{A}{The adjacency matrix that shows us the structure of the DAG. The input should be a lower triangular matrix where the possible matrix entry can only be 0}
}
\value{
The plot of the Directed Acyclic Graph that is described within the adjacency matrix
}
\description{
AdjDAG
}
\examples{
AdjDAG(A_matrix)
}
