% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefMat_Create_Spike_Slab.R
\name{CoefMat_Create_Spike_Slab}
\alias{CoefMat_Create_Spike_Slab}
\title{CoefMat_Create_Spike_Slab}
\usage{
CoefMat_Create_Spike_Slab(Y, A)
}
\arguments{
\item{Y}{The data the causal discovery adjacency matrix wants to show the connections with.}

\item{A}{The adjacency matrix that shows us the structure of the DAG. The input should be a lower triangular matrix where the possible matrix entry can only be 0}
}
\value{
A matrix that contains all the coefficients corresponding to specific relationships between each entry of the Y data
}
\description{
Uses spike and slab prior to get the corresponding coefficient matrix Bij corresponding to \eqn{Y = B_{ij}X_{i} + e_{i}}. The coefficients represents the connections from the jth entry of Y to the ith entry of Y.
}
\examples{
CoefMat_Create_Spike_Slab(Y_vector,A_matrix)
}
